Split update framework implementation
