+++ libmamba/src/validation/update_framework_v0_6.cpp
@@ -257,7 +257,7 @@ namespace mamba::validation::v0_6
 
     void to_json(nlohmann::json& j, const RootImpl& r)
     {
-        to_json(j, static_cast<const RoleBase*>(&r));
+        to_json(j, static_cast<const RoleBase&>(r));
     }
 
     void from_json(const nlohmann::json& j, RootImpl& role)
@@ -265,7 +265,7 @@ namespace mamba::validation::v0_6
         auto j_signed = j.at("signed");
         try
         {
-            from_json(j_signed, static_cast<RoleBase*>(&role));
+            from_json(j_signed, static_cast<RoleBase&>(role));
 
             role.set_timestamp(j_signed.at("timestamp").get<std::string>());
 
@@ -427,7 +427,7 @@ namespace mamba::validation::v0_6
 
     void to_json(nlohmann::json& j, const KeyMgrRole& r)
     {
-        to_json(j, static_cast<const RoleBase*>(&r));
+        to_json(j, static_cast<const RoleBase&>(r));
     }
 
     void from_json(const nlohmann::json& j, KeyMgrRole& role)
@@ -435,7 +435,7 @@ namespace mamba::validation::v0_6
         auto j_signed = j.at("signed");
         try
         {
-            from_json(j_signed, static_cast<RoleBase*>(&role));
+            from_json(j_signed, static_cast<RoleBase&>(role));
 
             role.set_timestamp(j_signed.at("timestamp").get<std::string>());
 
@@ -526,7 +526,7 @@ namespace mamba::validation::v0_6
 
     void to_json(nlohmann::json& j, const PkgMgrRole& r)
     {
-        to_json(j, static_cast<const RoleBase*>(&r));
+        to_json(j, static_cast<const RoleBase&>(r));
     }
 
     void from_json(const nlohmann::json& j, PkgMgrRole& role)
@@ -534,7 +534,7 @@ namespace mamba::validation::v0_6
         auto j_signed = j.at("signed");
         try
         {
-            from_json(j_signed, static_cast<RoleBase*>(&role));
+            from_json(j_signed, static_cast<RoleBase&>(role));
 
             role.set_timestamp(j_signed.at("timestamp").get<std::string>());
